Implemented a type safety policy to allow type safe usage of repositories' find methods. Three settings are supported:
NONE: No type safety is enforced on repository level. Thus the behaviour is the same as in past release. This is the default if the user selects no other policy via the Spring application context configuration.

RETURNS_NULL: If an entity with a given ID is loaded, but is from a different type as requested null is returned.

THROWS_EXCEPTION: If an entity with a given ID is loaded, but is from a different type as requested a InvalidEntityTypeException is thrown.
The policy can be selected by defining a Spring bean which is then autowired in the Neo config.

The policy is set by providing a bean with ID "typeSafetyPolicy" in the Spring application context configuration.

See http://stackoverflow.com/questions/13363607/resolving-entities-with-spring-data-neo4j-returns-wrong-entity-types/13375463#13375463 for a discussion concerning this.