DATAGRAPH-466 - Derived finders issue
Converted HashSet to TreeSet to maintain ordering as, without it, test
comparisons fail as the ordering of the match clause was unpredictable.
Although there may be a slight overhead in maintaining the order and
there's no real need to have a particular order, the predictability of
order is probably more valuable.