DATAGRAPH-511 - Tweak AspectJ setup to make sure we can build on Spring 4.1
Added aop.xml to only compile explicitly listed aspects into the code.  This is needed as Spring 4.1 includes a new aspect for JavaEE 7 JCache support that has optional dependencies which we don't have in the classpath. Trying to compile all aspects contained in spring-aspects will result in ClassNotFoundExceptions for the aspects with missing dependencies.