DATAGRAPH-602 - Incorrect propagation of @Transactional
This commit fixes a number of issues concerning transaction boundaries
in our Spring support classes. Specifically, propagation=REQUIRES was not
being joined to an existing transaction but always generated a new
one. Additionally, we were eagerly calling commit because the
implementation of isNewTransaction was incorrect. As a result partial
commits could be persisted, even if a subsequent commit failed, and
the whole transaction should have been rolled back.

This fix also provides support for REQUIRES_NEW, which was previously
missing.

Original pull request:#276.