DATAGRAPH-833 - Support @TransactionalEventListener.
The Neo4jTransactionManager needed to be overhauled. It now extends
AbstractPlatformTransactionManager rather than simply implementing
PlatformTransactionManager. This is because the current thread that a
session runs in isn't exposed further than what is configured for the
transaction manager itself.  We introduce the open session in view filter
(for those that prefer that method) and interceptor (the recommended way)
to bind the session to the current thread in a web context.

These changes remove the need for the session object to scope itself.

We also need to proxy the creation of the session so it is obtained from
the LocalSessionFactoryBean and not created outside that context. The
need to configure a Neo4jConfiguration is reduced.

We should consider wiring up sessions similar to how the JPA implementation
of spring data works. We may need Olivier to explain how that works though
as it's very cryptic.