DATAGRAPH-1061 - Register AuditingEventListener in post processor.
When using SpringBoot it will register all event listener in the
context it can find in the SessionFactory. This will also include
the Neo4jAuditionEventListener that would like to register itself.
This leads to a dependency cycle in the bean creation.
To get around this problem the registration will still be done in
SpringBoot but the self-registration will be removed from the
listener. If the listener is in the context, it will be registered
with a previous de-registration in a new BeanPostProcessor.

(cherry picked from commit caf0f5c)