DATAGRAPH-1061 - Avoid bean registration for AuditingEventListener.
By not registering the event listener as a bean in the context it is
cleaner to register it as an listener in the SessionFactory. There
is no need to de-register and register it again because it would not
be picked up by the SpringBoot auto configuration and registered there.

Additional the Neo4jAuditingBeanPostProcessor will only be instantiated
if the Neo4jAuditing is enabled.