DATAGRAPH-1068 - Respect sessionFactoryRef definition.
If sessionFactoryRef is defined in EnableNeo4jRepositories look for
a bean with the same name as the value of this property.
Previously there was no way to change the name ('sessionFactory') of
the required SessionFactory because the property was ignored.

(cherry picked from commit 5c541f2ddd5d8d69e6a983d7adc61e81a3607a6b)
with custom additions