DATAGRAPH-1094 - Support multiple SessionFactories
This change introduces full support for having multiple @EnableNeo4jRepositories-annotations using different instances of SessionFactories. There’s now an injectable shared Session for each SessionFactory qualified with the name of the corresponding SessionFactory.

There’s a subtle change in behavior that can break things in the case if a user already used the previous injectable session with the qualifier „sharedSessionCreatorBean“.