DATAGRAPH-1109 - Use OgmEntityInstantiatorAdapter for QueryResults, too.
The MappingContext has to be passed on, starting from the Neo4jRepositoryFactoryBean to the Neo4jRepositoryFactory and eventually to the GraphQueryMethod.

The GraphRepositoryQuery than uses it for creating the CustomResultConverter.
The OgmReflectionEntityInstantiator has been renamed and repurposed to a general QueryResultInstantiator which uses the OgmEntityInstantiator adapter if possible, otherwise the existing reflection based instantiator.