DATAGRAPH-1115 - Support @UseBookmark in composed annotations.
The annotation is not found if it is part of a composed annotation
when using a simple Class#getAnnotation call.
This changes the annotation access to use Springs AnnotatedElementUtils.

During the testing it was noticed that tests are only found name "*Tests.java",
so unrelated class Neo4jPersistentPropertyTest.java has been renamed
Neo4jPersistentPropertyTests.java.

# Conflicts:
#	spring-data-neo4j/src/test/java/org/springframework/data/neo4j/mapping/Neo4jPersistentPropertyTests.java