DATAGRAPH-1144 - Recognize not only internal ids as id properties.
Fix the determination of Spring Data Id Properties:

Before the fix, Neo4jPersistentProperty only considered the „identityField“ from OGMs ClassInfo as id property. This identifyField maps only to Neo4j internal, native ids.

As soon as a user annotates a non-long field with Id Spring Data cannot determine a required id property for the entity anymore.
While the entity can still be found by the „external“ id due to the fact that OGM keeps a cache mapping from external to internal / native, modules depending on knowing the property attribute, break.

The fix introduces a property type, which can be regular, internal id or id property.

Id properties are preferred over internal id properties based on the identityField as SDN id propertiers. This has the byproduckt that we now can have an „external“ id (the property mapped with @Id) and a „Long id“ field in an entity so that a developer can refer to both.