DATAGRAPH-1171 - Provide native Point to Spring Data Commons Point converter.
This adds an internal converter to each field of
org.springframework.data.geo.Point that converts Springs Point to a
GeographicPoint2d before passing it on to OGM when OGMs support for
native types is enabled.

We choose a geographic point and not a cartesian one because SDN mapped
x/y to latitude/longitude. As a matter of fact, this implicit conversion
maps x to latitude and y to longitude in contrast of OGMs and Neo4js
mapping, to keep the behaviour consistent with non-native spatial types.

The commit also adds a customized query for "Near" query: The statement
now checks wether a location of a domain object has been written using
non-native types (as two properties on a node) or as a native point and
chooses them in that order.

Last but not least, this change also contains the fix of DATAGRAPH-1203
with corresponding tests.