Create NodeManagerFactory, NodeManager and related infrastructure.
A NodeManager will be used to create, read, update, delete and query managed objects.
It will use a short lived persistence context.

This adds the basic skeletons for doing this but especially the infrastructure to allow the following usage:
* User configures a driver bean
* User provides a NodeManagerFactory using that driver and a given set of classes to be managed
* The factory gets augmented to handle transactions and facilitate Spring Datas class scanning mechanism

The Spring Boot starter will automatically provide a configured NodeManager instance.