Introduce JUnit 5 tags for various types of tests.
This introduces some constants to mark tests as tests that need reactive
support and to distinguish between tests that run only on the commercial
or community edition.

The Neo4jExtension assumes on the requirements of the tags, so those
tags have a dual use:

They can be used to filter for tests, i.e. setup a pipeline that does
only reactive tests on community edition.

They make sure the project is easy to build for external contributors
and people who might not have access to latest edition by assuming on
the requirements and prevent the tests from failing in conditions where
the tests are bound to fail.