Add support for CustomConversions.
This adds `Neo4jConversions`, an implementation of Spring Datas `CustomConversions`. A default instance is provided through the configuration support and also in the automatic configuration of Spring Boot.

The `Neo4jConversions` is than used to create an internal `ConversionService` used by a `Neo4jConverter`, that provides methods for reading and writing values and also for decorating property accessors and parameter value providers.

While writing the conversion, it turned out that we could postpone converting from drivers `Value` as late as possible.

Thus all possible user provided conversions need to be conversions from `org.neo4j.driver.Value` to the desired type and back.

This closes #23.