Add relationship properties mapping.
Provides the functionality of defining relationship with
additional property classes represented as a map structure.

Also the mapping and binder functions for loading or saving data
were condensed into central anonymous functions and shifted
their mapping logic to the Neo4jConverter.

Required behaviour changes in the existing code base:
For setting the Driver bean the Neo4jMappingContext must not be an implementation
of a BeanDefinitionRegistryPostProcessor because this would
require setting up the driver bean very early in the start process
that e.g. configuration parameters that also rely on PostProcessor logic
cannot be processed before setting up the driver.