GH-128 - Make database selection reactive capable.
While the database name provider in fd8068b24e5a80fc1ded6e5a884ddd6447de9c02 works, it doesn’t help in some reactive scenarios, for example, with the reactive security principle.

We need to role an independent version of it, returning the database selection as a Mono.

While doing this, it becomes obvious that we shouldn’t deal with the raw database name. The value of null is neither a valid database name and nor a valid mono value.
Therefor the notion of a database selection is introduced, with an explicit value of „undecided“ for all the case in which the server should decide.