Make database selection reactive capable.
While the database name provider in 505d9658823a82c259271e25467705eb5e304a1a works, it doesn’t help in some reactive scenarios, for example, with the reactive security principle.

We need to role an independent version of it, returning the database selection as a Mono.

While doing this, it becomes obvious that we shouldn’t deal with the raw database name. The value of null is neither a valid database name and nor a valid mono value.
Therefor the notion of a database selection is introduced, with an explicit value of „undecided“ for all the case in which the server should decide.