Make Spring Boot autoconfiguration aware of custom EnableNeo4jRepositories.
This makes both the imperative and reactive variants of the autoconfiguration aware of existing, custom `@EnableNeo4jRepositories` respectively `@EnableReactiveNeo4jRepositories` by adding a `@ConditionalOnMissingBean` checking for the presence of the corresponding factories.

`DataNeo4jTest` needs to be removed from the reactive test. It is problematic when configured with only reactive repositories in use as it automatically enables transaction management. The correct choice with `spring.data.neo4j.repositories.type=reactive` is the combination of `@ExtendWith(SpringExtension.class)` and  `@AutoConfigureDataNeo4j` or, when run without our test support, a full `@SpringBootTest`.

This fixes #133.