GH-159 - Use Neo4j's causal bookmarks.
This brings in support for causal bookmarks of Neo4j. It requires no configuration from the user nor an additional library. It works out of the box for everything that runs inside a Spring managed transaction, for both imperative and reactive scenarios.

The core idea for both scenarios is as follows:

* Keep a list of all bookmarks ever retrieved from a cluster
* Pass that list to new transactions, the server sorts and merges them for us, no need to worry about the order
* When the transaction completes, collect the last bookmark from it and replace the used bookmarks from beforehand (and only them) with it. Thus, the bookmark returned by a parallel transactions is not affected
* The bookmarks work as a "at least" indicator, so all new transactions will see at least what they have written and are thus serialisable on the application side of things.

This closes #158. 

Co-authored-by: Gerrit Meier <meistermeier@gmail.com>