Support path pattern in conditions.
This change adds additional overloads for `where` and `and`, `or` and `xor` in the `StatementBuilder` and on the `Condition` interfaces.

These allow a `RelationshipPattern` to be used in condition as described in the Neo4j 4.0 docs chapter 3.6.5 and closes #206.

The `RelationshipPattern` corresponds directly with the "path pattern" mentioned in the docs. The name is taken from the OpenCypher spec.

In addition to this change, the `ExposesRelationships` has been moved to the `RelationshipPattern` to allow for easier chaining.

Furthermore, it was noted that in the Spring Data related parts, `org.neo4j.springframework.data.repository.query.CypherQueryCreator#complete` did not cater for the fact that the last - or only condition - might be null.