Add support for list comprehensions.
This commit brings support for generating list comprehensions. It also optimizes the operator for generating pattern comprehension in regards of replacing the `Pipe` literal with an `Operator.PIPE`, resulting in nicer renderings.
The `Operation` expression excludes this `PIPE` (and the existing `EXPONENTIATION` operator) from groups.

The `PatternComprehension` expression has been reworked to only expose correct builders, in accordance with the new `ListComprehension`.

This commit closes #225.