GH-231 - Add support for dynamic labels.
This adds support for dynamic or runtime managed labels. Attributes of type `Collection<String>` can be annotated with `@DynamicLabels`. During reads of those entities, all labels that don't fit on statically defined labels via the class hierarchy, will be added to that collection. During save, all statically defined labels plus the ones in the dynamic labels collection will be used for a node. All other labels will be deleted.

This change doesn't require a state holder in SDN/RX. If `@DynamicLabels` is used, the "truth" about the labels of a node is defined by the entity being save.

This closes #214.

Co-authored-by: Gerrit Meier <meistermeier@gmail.com>