Allow different relationship types to the same entity instance.
This allows for multiple relationships of different types between the same two objects. It is implemented via a simple state machine that checks both the type of the relationship and the state of the related objects already processed before terminating the traversal.

In addition the change makes sure that only the object returned by Spring Data Commons event support is passed on in the recursive traversal.

Last but not least, we switch to Neo4j 4.0 in tests by default to make testing of reactive components active by default.