DATAGRAPH-1373 - Improve RelationshipProperties API.
The relationship with properties will now define the `TargetNode`
that points to the entity that should be referred to, either outgoing or
incoming.
The syntax with the `Map` usage was discarded because it has
limitations in the convinience usage of the derived finders and did not
allow the definition of mulitple relationships of the same type to the
same node.