DATAGRAPH-1212 - Improve fallback IsNewStrategy.
This change replaces the previous fallback IsNewStrategy with a backport from SDN/RX respectivley SDN 6. This takes much better care of handling externally assigned and primitive ids together with version properties and fails fast when it can’t cater for a scenario.

In addition to the required OGM update, this makes it possible to audit entities with externally generated or assigned ids. For this scenario, tests have been added as well.