DATAGRAPH-1386 - Log Query notifications.
Neo4j provides notifications for queries that might have triggered a deprecated feature. We log them on `cypherLog` and make them visible to the user in all queries going through the Neo4j clients.

The notifications come with a severity of `WARNING` and `INFORMATION` which is used for selecting the appropriate level.