DATAGRAPH-1387 - Finalize and document package structure.
This moves internal api around to better reflect the intent of the public API by making it stand out on its own. A rule has been added to make sure public api packages are not directly dependent on the mapping code.

The public package structure has been documented in the building blocks.