DATAGRAPH-1393 - Add support for OGM's CompositeAttributeConverter.
This brings in the necessary infrastructure for decomposing single attributes of an entity into a map of properties that are all written to the node or relationship.

The `CompositeProperty`-annotation lives in the `schema` package as it defines the schema from the point of the Java mapping code.

`CompositeProperty` will work out of the box on maps with key of types string or enum, with configurable prefix and separator. In addition with additional converters it can be applied to attributes of custom types, too.