DATAGRAPH-1397 - Create correct target collections for relationships.
The target type of the collection for relationships with properties has not been taken into account and this commit fixes it.

The commit also addresses a possible issue in the Cypher generator for relationships when several mapping to different nodes with the same type exists. The loop needs to continue in case one has already been found, not left.

 Co-authored-by: Gerrit Meier <meistermeier@gmail.com>