DATAGRAPH-1401 - Update outdated chapter about testing.
This removes the references to the SDNRX test autoconfiguration and provides 3 viable and supported alternatives for using `@DataNeo4jTest` with Test harness or Test containers.

It also takes the changes in Spring Boot 2.4 into consideration in which `@DataNeo4jTest` is no longer `@Transactional`.

The commit removes all notion of `@ReactiveDataNeo4jTest`.