DATAGRAPH-1429 - Allow the use of one top-level aggregate.
We have everything in place to call our entity reader several times. The biggest change here is to pass in the map accessor from the outside, not the record, which I should have done when I changed the entity reader to work on the map accessor in the first place.

The prepared query now has a mapping function that knows if it had to unwrap a list before mapping and aggregating it back afterwards. If a list had been unwrapped, then allResults, both the imperative and reactive ones, will react accordingly.

This change allows the flight scenario and a couple of other things to work.

Custom maps can be returned as well, by using DTO based projections with additional properties (returning them as `[{node: n, p1: "p1", p1: "pn“}]` and than applying a custom streamable.