DATAGRAPH-1441 - Register necessary callbacks internally inside the mapping context.
IdGenerating callbacks must be registered for each Neo4j mapping context, otherwise
they could check with the wrong context whether an entity is new or requires
optimistic locking. In a case where the context is strict, this will lead to an
exception.

The callbacks are now registered in the templates belonging to the context.
With this change, we aren’t dependend on the correct order of callback registrars
anymore.
It is next to impossible to find a way that deals with multiple context that works
both in a plain Spring scenario only using our configuration,
the default Spring Boot scenario or a completly custom one.
There are too many ways a user can end up missing those required beans.

In the light of that change it became appearant that the event / callback system
doesn't belong inside the repository package but in the mapping one. As the id
generating and versioning callbacks have not been public API in the first place,
they have been moved directly to the package without a previous deprecation.
They are not package private and only accessible through internal API.

The auditing callbacks on the other hand might be instantiated directly by a user
that doesn't want to use the dedicated annotation or the additional registrar.
Therefore they have been deprecated and they delegate now to the replacement.

The general before bind callback interfaces - both imperative and reactive - have
been deprecated as well and replaced by a stable API in the new package. In 6.0.2
upwards, both will be used.
They will be removed in 6.1.x.

Last but not least, the registrars bringing in the id generating and versioning
callbacks are now no-op operations and deprecated. SDN doesn't need them anymore.

While working on the support for multiple context it became appearant that is sensible
to extract the class scanning and provide it as a utility method rather than
support class.