Always register the required callbacks in the templates.
With this change, we aren’t dependend on the correct order of callback registrars (custom `ImportBeanDefinitionRegistrar` implementations anymore). It is next to impossible to find a way that deals with multiple context that works in a plain Spring scenario only using our configuration, the default Spring Boot scenario or a completly custom one. There are too many ways a user can end up missing those required beans.