DATAGRAPH-1449 - Delete removed dynamic relationships with properties.
The Cypher generator always looked at the entity describing the related entity - which is the properties holder in the case of relationship with properties - when computing which relation to delete.

The fix for DATAGRAPH-1449 is to use the relationship description. In addition, there was a slight chance that a wrong ID converter was used.

Tests have been adapter to force a full refresh of the entities.