DATAGRAPH-1440 - Improve support for methods returning Slice.
This aligns SDN 6 with the other Spring Data module that a count query is
optional for queries that return a `Slice<T>` and mandatory for queries that
return a `Page<T>`. SDN will fail paged methods with custom queries  that
are missing a count query and will query for one more item on sliced methods without
a dedicaded count query.

For that a flag has been introduced if the limit should be raised. To avoid modifying
the `Pageable` object in place, an operator todo so is passed along to the Cypher
generator.

In addition to that change SDN will log warnings in case

- A method returning a page or a slice with a custom query was called with a sorted `Pageable`.
- A method returning with a custom query was called with a sorted `Sort`.
- A method returning a page or a slice with a custom query doesn't provide the two
  Cypher parameters `$skip` and `$limit`