DATAGRAPH-1454 - Add support for literal SpEL replacements.
This change adds support for two Neo4j custom SpEL extensions:

- `literal()`: Replaces the SpEL fragment with the literal value returned by the expression
   instead of defining an additional Cypher parameter
- `orderBy()`: Takes in a `Pageable` or `Sort` object and creates `ORDER BY` fragment.

For this to work, `Neo4jSpelSupport` and `Neo4jEvaluationContextExtension` has
been added. For the moment the extension is used in both imperative and reactive
implementations, as the support for reactive SpEL evaluation is not ready yet in SDN.

Both imperative and reactive String based queries have been changed to recognized
literal replacements of expression and the shared code has been moved into the
the SpEL-support.

The orderBy extension won't check if the attributes returned actually are present
in the query or not. This is the users task.