GH-2099 - Declare the dependency between shared session creator and entity instantiator.
The conversion service is picked up to late as the the dependency
between the shared session creator and the bean that does the post
processing of the underlying session factory is undefined for the Spring
context.

By explicit stating this in the repository extension, the
`Neo4jOgmEntityInstantiatorConfigurationBean` is correctly initialized
upfront without messing around with an actual post processor or
orderings.

This closes #2099.