GH-1691 - Fail early if bookmark setup is enabled without a bookmark manager.
This change adds a bean through `@EnableBookmarkManagement` that asserts
the presence of a bean implementing
`org.springframework.data.neo4j.bookmark.BookmarkManager`. Thus the
context will fail to start up early when the user assumes that the
combination of `@EnableBookmarkManagement` and `@UseBookmark` is enough
and not only on the latest point in time possible.

This closes #1691.