GH-2114 - Include all nodes of a path during mapping.
This changes the logic of path mapping. We now synthesize a result that has the form as a custom query would have (root node, all relationships and all related nodes), instead of going through the paths. This simplifies the entity converter by a magnitude.

Thus, we automatically include all nodes of a given type when we are not interested in the relation in a query like
`MATCH p=()-[:IS_SIBLING_OF]-> () RETURN p`, include the last node in the path as an related object, are independent of both the direction that has been used for querying when filling in relationships and also have the same results regardless whether the paths is odd or even numbered.

This fixes #2114.