GH-2123 - Select correct conversion for assigned ids of related nodes.
We must select the source entity and it's associcated id property for
correctly picking up the right conversion as the source is only matched
by the assigned id.

It does not matter for new related target noode as that one is created
or merged in the same step and we can work with its internal id.

This fixes #2123.