GH-2133 - Remove `o.s.d.annotation.Persistent` from the set of identifying annotations.
This removes `@Persistent` from the set that is used as so called
identifying annotations inside the repository configuration.

This change makes the process of identifying a repository as Neo4j
repository more strict: While it doesn't change anything in a single
store scenario - repository is selected by the entity being managed and
that is still `@Persistent` - it will fix missidentifying a repository
of another store as elligable for being a Neo4j repository.

This fixes #2133.