GH-1911 - Polishing.
This changes the return types of the `CypherdslConditionExecutor` into `Collection` and aligns it with the `CypherdslStatementExecutor` and in the long run with the goal Spring Data Commons has for the `QuerydslPredicateExecutor`.  It makes more sense for this new API to align with the desired state instead of something we consider suboptimal.