GH-2168 - Treat no attributes annotated with `@ConvertedWith` as association.
The actuall issue here is caused by `@ConvertedWith` not indicating at
all times that a property isn't an association: The moment there is a
converter (either through the Spring context or `@ConvertedWith`) it
can't be an association anymore.

The exception `Required identifier property not found for class` was
thrown because SDN treated the annotated property as association. This
is fixed now.

The commit adds tests for this and also adds an example how to use
composite or discrete property convertes. However, in alignment with
previous SDN+OGM behavior, a prefixless decomposition of properties into
maps and composition back from them is not possible.

This closes #2168.