GH-2182 - Add support for projections inside the Neo4j templates.
This adds support for projections inside Neo4j templates in a couple of places.
All changes are applied into the reactive and the imperative world:

- `saveAs`
- `saveAllAs`

Methods have been added directly to the `Neo4jOperations`. Those take in a class or interface that should be returned.
The same logic as with projections during reads is applied: Only properties and relationships that are part of the projection are save / updated. All others are left alone. This works exactly one level deep as of now and might change in the future with enhanced Spring Data Commons support. Those methods have been defaulted so that they don’t break peoples implementation of the operations.

In addition, `FluentNeo4jOperations` and some support classes have been introduced that allow to specify the project for finder methods.
The default templates implement does now.

Both templates bring now their own `ProjectionFactory`.

This closes #2182.