GH-2207 - Improve the pair of isEntity and isAssociation.
This introduces the idea of a "writable property": Something that can be
written "as is" to the database. This is a rather complicated decision
to make due to the constructs we support back from Neo4j-OGM days.

It goes hand in hand with the decision whether a given property on a
entity of the domain describes an association: Everything that is either
explicitly annotated with `@Relationship` or is not something that would
appear as a graph property is an association.

A property on the domain model on the other hand is an entity when it is
not something that can be written directly to the graph.

That leaves the special case of entities with properties: Those life
quite literally between entities, as associations with properties. To
make them work in the context of Spring Data Commons, they must be
treated as entities and in a dedicated way in our converters. This is
now encapsulated in the (already) existing dedicated method
`isEntityWithRelationshipProperties()`. However, this method is no longer
called from `isEntity()` to solve the cyclic meaning here.